# openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "タビキメワリ Business API",
    "version": "0.1.0",
    "description": "M0-A3 business API contract. Authentication, invite, internal task, upload transport, and event contracts are completed with the B lane."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "security": [
    {
      "firebaseBearer": []
    }
  ],
...
  "paths": {
...
    "/api/trips": {
      "get": {
        "operationId": "listTrips",
        "tags": [
          "Trips"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/Cursor"
          },
          {
            "$ref": "#/components/parameters/PageSize"
          }
        ],
        "responses": {
          "200": {
            "description": "ACTIVE membership trips",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripPage"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      },
      "post": {
        "operationId": "createTrip",
        "tags": [
          "Trips"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTripRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Trip, OWNER member, and initial slots created atomically",
            "headers": {
              "Location": {
                "$ref": "#/components/headers/Location"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripSnapshot"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/ValidationFailed"
          }
        }
      }
    },
    "/api/trips/{tripId}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/TripId"
        }
      ],
      "get": {
        "operationId": "getTripSnapshot",
        "tags": [
          "Trips"
        ],
        "responses": {
          "200": {
            "description": "Current trip snapshot",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripSnapshot"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "patch": {
        "operationId": "updateTrip",
        "tags": [
          "Trips"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTripRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated trip",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Trip"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/ValidationFailed"
          }
        }
      }
    },
...
    "/api/trips/{tripId}/expenses": {
      "parameters": [
        {
          "$ref": "#/components/parameters/TripId"
        }
      ],
      "get": {
        "operationId": "listExpenses",
        "tags": [
          "Expenses"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/Cursor"
          },
          {
            "$ref": "#/components/parameters/PageSize"
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/ExpenseStatus"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Cursor-paged expenses",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExpensePage"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "post": {
        "operationId": "createExpenseDraft",
        "tags": [
          "Expenses"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateExpenseDraftRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Expense DRAFT created",
            "headers": {
              "Location": {
                "$ref": "#/components/headers/Location"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Expense"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/ValidationFailed"
          }
        }
      }
    },
    "/api/trips/{tripId}/expenses/share-preset": {
      "parameters": [
        {
          "$ref": "#/components/parameters/TripId"
        }
      ],
      "get": {
        "operationId": "getPreviousExpenseSharePreset",
        "tags": [
          "Expenses"
        ],
        "responses": {
          "200": {
            "description": "Share inputs copied from the latest confirmed expense",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExpenseSharePreset"
                }
              }
            }
          },
          "204": {
            "description": "No confirmed expense is available for a previous-share preset"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/api/trips/{tripId}/expenses/{expenseId}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/TripId"
        },
        {
          "$ref": "#/components/parameters/ExpenseId"
        }
      ],
      "get": {
        "operationId": "getExpense",
        "tags": [
          "Expenses"
        ],
        "responses": {
          "200": {
            "description": "Expense detail",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Expense"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "patch": {
        "operationId": "updateExpense",
        "tags": [
          "Expenses"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateExpenseRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated or confirmed expense",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Expense"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/ValidationFailed"
          }
        }
      },
      "delete": {
        "operationId": "deleteExpenseDraft",
        "tags": [
          "Expenses"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/Version"
          }
        ],
        "responses": {
          "204": {
            "description": "DRAFT deleted; CONFIRMED expenses require an audited correction"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          }
        }
      }
    },
    "/api/trips/{tripId}/expenses/{expenseId}/receipt-upload": {
      "parameters": [
        {
          "$ref": "#/components/parameters/TripId"
        },
        {
          "$ref": "#/components/parameters/ExpenseId"
        }
      ],
      "post": {
        "operationId": "prepareExpenseReceiptUpload",
        "tags": [
          "Expenses"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PrepareReceiptUploadRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Short-lived upload capability. The URL must never be logged.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReceiptUpload"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/ValidationFailed"
          }
        }
      }
    },
    "/api/trips/{tripId}/expenses/{expenseId}/receipts/{receiptId}/completion": {
      "parameters": [
        {
          "$ref": "#/components/parameters/TripId"
        },
        {
          "$ref": "#/components/parameters/ExpenseId"
        },
        {
          "$ref": "#/components/parameters/ReceiptId"
        }
      ],
      "post": {
        "operationId": "completeExpenseReceiptUpload",
        "tags": [
          "Expenses"
        ],
        "description": "Verifies the stored object's size and content type before marking the receipt UPLOADED.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VersionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Expense with verified uploaded receipt",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Expense"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/ValidationFailed"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "firebaseBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Firebase ID token",
        "description": "Signature-verified Firebase ID token. Travel roles are loaded from trip_member, never trusted from client claims."
      },
      "tasksServiceOidc": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Google service-account OIDC token",
        "description": "Only the configured Cloud Tasks service account and exact environment audience are accepted."
      },
      "schedulerServiceOidc": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Google service-account OIDC token",
        "description": "Only the configured Cloud Scheduler service account and exact environment audience are accepted."
      }
    },
    "parameters": {
      "TripId": {
        "name": "tripId",
        "in": "path",
        "required": true,
        "schema": {
          "$ref": "#/components/schemas/Id"
        }
      },
...
      "ExpenseId": {
        "name": "expenseId",
        "in": "path",
        "required": true,
        "schema": {
          "$ref": "#/components/schemas/Id"
        }
      },
...
      "Cursor": {
        "name": "cursor",
        "in": "query",
        "required": false,
        "schema": {
          "type": "string",
          "maxLength": 500
        },
        "description": "Opaque server-issued cursor"
      },
      "PageSize": {
        "name": "limit",
        "in": "query",
        "required": false,
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 100,
          "default": 20
        }
      },
      "IdempotencyKey": {
        "name": "Idempotency-Key",
        "in": "header",
        "required": true,
        "schema": {
          "type": "string",
          "format": "uuid"
        },
        "description": "Reused for retries of the same logical operation. A reused key with a different request hash returns 409."
      },
      "Version": {
        "name": "version",
        "in": "query",
        "required": true,
        "schema": {
          "$ref": "#/components/schemas/Version"
        }
      }
    },
    "headers": {
      "Location": {
        "description": "Canonical URL of the created resource",
        "schema": {
          "type": "string",
          "format": "uri-reference"
        }
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Malformed request or idempotency header",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Missing or invalid Firebase identity",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Authenticated trip member lacks the required role",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "NotFound": {
        "description": "Trip membership or nested resource ownership was not found; this response does not disclose another trip's resource",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "Conflict": {
        "description": "Optimistic-lock, idempotency, or domain-state conflict",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ConflictProblem"
            }
          }
        }
      },
      "ValidationFailed": {
        "description": "Syntactically valid request that violates field or business validation",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
...
    },
    "schemas": {
...
      "Trip": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "title",
          "destination",
          "startsOn",
          "endsOn",
          "timezone",
          "expectedMemberCount",
          "phase",
          "voteVisibility",
          "revision",
          "version"
        ],
        "properties": {
          "id": {
            "$ref": "#/components/schemas/Id"
          },
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "destination": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "startsOn": {
            "type": "string",
            "format": "date"
          },
          "endsOn": {
            "type": "string",
            "format": "date"
          },
          "timezone": {
            "type": "string",
            "minLength": 1,
            "maxLength": 40,
            "description": "Validated IANA timezone"
          },
          "expectedMemberCount": {
            "type": "integer",
            "minimum": 1,
            "maximum": 100
          },
          "phase": {
            "$ref": "#/components/schemas/TripPhase"
          },
          "phaseOverride": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/TripPhase"
              },
              {
                "type": "null"
              }
            ]
          },
          "voteVisibility": {
            "$ref": "#/components/schemas/VoteVisibility"
          },
          "budgetCap": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/Yen"
              },
              {
                "type": "null"
              }
            ]
          },
          "revision": {
            "$ref": "#/components/schemas/Version"
          },
          "version": {
            "$ref": "#/components/schemas/Version"
          }
        }
      },
      "CreateTripRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "title",
          "destination",
          "startsOn",
          "endsOn",
          "timezone",
          "expectedMemberCount",
          "ownerName"
        ],
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "destination": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "startsOn": {
            "type": "string",
            "format": "date"
          },
          "endsOn": {
            "type": "string",
            "format": "date"
          },
          "timezone": {
            "type": "string",
            "minLength": 1,
            "maxLength": 40
          },
          "expectedMemberCount": {
            "type": "integer",
            "minimum": 1,
            "maximum": 100
          },
          "ownerName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          },
          "budgetCap": {
            "$ref": "#/components/schemas/Yen"
          },
          "voteVisibility": {
            "$ref": "#/components/schemas/VoteVisibility"
          }
        }
      },
...
      "UpdateTripRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "version"
        ],
        "minProperties": 2,
        "properties": {
          "version": {
            "$ref": "#/components/schemas/Version"
          },
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "destination": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "startsOn": {
            "type": "string",
            "format": "date"
          },
          "endsOn": {
            "type": "string",
            "format": "date"
          },
          "timezone": {
            "type": "string",
            "minLength": 1,
            "maxLength": 40
          },
          "expectedMemberCount": {
            "type": "integer",
            "minimum": 1,
            "maximum": 100
          },
          "phaseOverride": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/TripPhase"
              },
              {
                "type": "null"
              }
            ]
          },
          "voteVisibility": {
            "$ref": "#/components/schemas/VoteVisibility"
          },
          "budgetCap": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/Yen"
              },
              {
                "type": "null"
              }
            ]
          }
        }
      },
      "TripPage": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "items"
        ],
        "properties": {
          "items": {
            "type": "array",
            "maxItems": 100,
            "items": {
              "$ref": "#/components/schemas/Trip"
            }
          },
          "nextCursor": {
            "oneOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ]
          }
        }
      },
      "TripSnapshot": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "trip",
          "currentMemberId",
          "members",
          "slots",
          "planItems"
        ],
        "properties": {
          "trip": {
            "$ref": "#/components/schemas/Trip"
          },
          "currentMemberId": {
            "$ref": "#/components/schemas/Id"
          },
          "members": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Member"
            }
          },
          "slots": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Slot"
            }
          },
          "planItems": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PlanItem"
            }
          }
        }
      },
...
      "ExpenseStatus": {
        "type": "string",
        "enum": [
          "DRAFT",
          "CONFIRMED"
        ]
      },
...
      "Expense": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "createdByMemberId",
          "source",
          "status",
          "receipts",
          "shares",
          "version"
        ],
        "properties": {
          "id": {
            "$ref": "#/components/schemas/Id"
          },
          "planItemId": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/Id"
              },
              {
                "type": "null"
              }
            ]
          },
          "createdByMemberId": {
            "$ref": "#/components/schemas/Id"
          },
          "payerId": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/Id"
              },
              {
                "type": "null"
              }
            ]
          },
          "amount": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/PositiveYen"
              },
              {
                "type": "null"
              }
            ]
          },
          "currency": {
            "oneOf": [
              {
                "type": "string",
                "const": "JPY"
              },
              {
                "type": "null"
              }
            ]
          },
          "baseAmount": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/PositiveYen"
              },
              {
                "type": "null"
              }
            ]
          },
          "paidAt": {
            "oneOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ]
          },
          "source": {
            "type": "string",
            "enum": [
              "PLAN",
              "MANUAL"
            ]
          },
          "status": {
            "$ref": "#/components/schemas/ExpenseStatus"
          },
          "allocationType": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/AllocationType"
              },
              {
                "type": "null"
              }
            ]
          },
          "receipts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Receipt"
            }
          },
          "shares": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ExpenseShare"
            }
          },
          "version": {
            "$ref": "#/components/schemas/Version"
          }
        }
      },
      "CreateExpenseDraftRequest": {
        "type": "object",
        "additionalProperties": false,
        "anyOf": [
          {
            "required": [
              "amount"
            ]
          },
          {
            "properties": {
              "hasReceipt": {
                "const": true
              }
            },
            "required": [
              "hasReceipt"
            ]
          }
        ],
        "properties": {
          "planItemId": {
            "$ref": "#/components/schemas/Id"
          },
          "amount": {
            "$ref": "#/components/schemas/PositiveYen"
          },
          "paidAt": {
            "type": "string",
            "format": "date-time"
          },
          "source": {
            "type": "string",
            "enum": [
              "PLAN",
              "MANUAL"
            ],
            "default": "MANUAL"
          },
          "hasReceipt": {
            "type": "boolean",
            "description": "Must be true when creating a photo-only DRAFT; an upload is prepared after creation"
          }
        }
      },
...
      "ExpensePage": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "items"
        ],
        "properties": {
          "items": {
            "type": "array",
            "maxItems": 100,
            "items": {
              "$ref": "#/components/schemas/Expense"
            }
          },
          "nextCursor": {
            "oneOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ]
          }
        }
      },
      "ExpenseSharePreset": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "sourceExpenseId",
          "allocationType",
          "shares"
        ],
        "properties": {
          "sourceExpenseId": {
            "$ref": "#/components/schemas/Id"
          },
          "allocationType": {
            "$ref": "#/components/schemas/AllocationType"
          },
          "shares": {
            "type": "array",
            "minItems": 1,
            "maxItems": 100,
            "items": {
              "$ref": "#/components/schemas/ExpenseShareInput"
            }
          }
        }
      },
      "PrepareReceiptUploadRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "contentType",
          "byteSize",
          "version"
        ],
        "properties": {
          "contentType": {
            "type": "string",
            "enum": [
              "image/jpeg",
              "image/png",
              "image/webp"
            ]
          },
          "byteSize": {
            "type": "integer",
            "minimum": 1,
            "maximum": 10485760
          },
          "version": {
            "$ref": "#/components/schemas/Version"
          }
        }
      },
...
    }
  }
}